fix: review

+++ api/src/main/kotlin/com/gotchai/api/presentation/v1/exam/response/SubmitExamResponse.kt
@@ -4,13 +4,13 @@ import com.gotchai.api.presentation.v1.badge.response.BadgeResponse
 import com.gotchai.domain.exam.dto.result.ExamSubmitResult
 
 data class SubmitExamResponse(
-    val answerCount: Long,
+    val answerCount: Int,
     val badge: BadgeResponse
 ) {
     companion object {
         fun from(submitExam: ExamSubmitResult): SubmitExamResponse =
             SubmitExamResponse(
-                answerCount = submitExam.correctAnswerCount.toLong(),
+                answerCount = submitExam.correctAnswerCount,
                 badge = BadgeResponse.from(submitExam.badge)
             )
     }
